requirements.txt
pandas>=2.2,<3
sentence-transformers>=3.0,<4
chromadb>=0.5,<0.7
gradio>=5.0,<6
python-dotenv>=1.0,<2
